requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==3.0.0
Pillow==10.0.0
fuzzywuzzy[speedup]==0.19.1
gunicorn==20.1.0
ImageHash==4.3.1